AI-Agents/Agent-04-openrouter/pyproject.toml
[project]
name = "agent-04-openrouter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
]
